Added containerd support
Context:

Since kubernetes 1.25 docker is being removed as the default container runtime in favor of containerd.

Due this is desired to have a mechanism to parse containerd auth files following the spec defined here:

https://github.com/containerd/containerd/blob/main/docs/cri/config.md#registry-configuration

On this PR is being implemented the necesary code to parse the containerd toml files and login on the registries defined on them.