Support vnd.oci.image.index.v1+json media type for choosing the architecture
During the debugging of this help center ticket, we have been raised with the concern that for an image built to support arm64 architectures, it was not being scheduled in the right nodes unless it was forced.

While doing some research on the information persisted on an image registry for this image, we've detected that the mediaType interface on which this manifest is served, is not being taken into account by noe's code, thus not being supported, and expecting values that are not in the structure served by this endpoint.

An example response is:

    ```

    {
        "mediaType": "application/vnd.oci.image.index.v1+json",
        "schemaVersion": 2,
        "manifests": [
            {
                "mediaType": "application/vnd.oci.image.manifest.v1+json",
                "digest": "test-test",
                "size": 676,
                "platform": {
                    "architecture": "arm64",
                    "os": "linux"
                }
            }
        ]
    }

    ```

Since noe expects to have a `platform` key value in the main object, and it's not specified here, the code will assume an assignment of a default value (amd64), and set the node selector accordingly.

This pull request is adding the support for this new media type, which will allow the right platform to be recognized. The current structure to unmarshal the response is working accordingly and does not require further changes.

Also, there is a serialized value for the platform labeled as "unknown", which is being filtered out to avoid issues.