Add logic to check and ignore long running pods (#55)
# Context

We have observed that when Noe started, it reconciled too many pods in
the cluster even if not entirely needed. This may be a non-issue at
small scales. However, reconciling too many pods at the same time
could generate too many requests to the image repository/ies resulting
in having 401 answers from every subsequent attempt making Noe fail to
function.

With this PR, we are excluding a number of pods that do not needs to
be reconciled when Noe restarts by only reconciling the required
ones. Noe already has a cache for images it already reconciled but
it's in-memory so it's emptied when Noe restarts.

 # What condition do we consider for skipping the reconciling

- If the pod is Running and having a Ready condition as True

This basically means that we will ignore these pods since we assume
they are placed into the right node with the right architecture, so
there's no need for noe to reconcile them.

From the documentation:
https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-readiness-gate

    A Pod's condition is Ready, when all of its containers are running
    properly and those with a readinessProbe are passing the probe. If any
    container in the Pod is not running or fails its readinessProbe, the
    Pod's Ready condition will be False.

In other words, when a Pod is in a Ready condition, you can be assured
that all its containers are running properly. If there were issues
with any container, the Pod would not be marked as Ready.

With this assumption, it means pods with Ready condition have been
running in the right node's architecture.

Change-Id: I47c8a28bc06fd996d96158e3a886282783e4e15e
Co-authored-by: Tanat Lokejaroenlarb <tanatloke@gmail.com>