bug: StatefulSets are not reevaluated after rescheduling
 # Context

Noe caches the Pods it has already processed to avoid reprocessing
them unless there are changes. We do this by storing the
`NamespacedName` of the Pod, which includes the namespace and Pod
name.

For Deployments, rescheduling the Pod actually changes its name as the
termination is generated. For StatefulSets, the name is kept the same
no matter how many reschedules happen.

Our caching approach prevents us form reevaluating the location of
StatefulSet pods if they were previously scheduled, and we've seen
this allocate StatefulSet pods to nodes whose architectures where not
supported by the Pod.

 # Bugfix approach

The intention of the cache is to reduce load over Registries and Noe
itself. In our case so far StatefulSets are much less common than
Deployments, so we choose to avoid the cache for StatefulSets.

This is a stopgap measure to prevent misscheduling of current
workloads until a better approach to prevent reprocessing of Pods can
be devised that supports StatefulSets.

 # What we change

We add a `skipCache(*v1.Pod)` function that checks if the Pod is owned
by a StatefulSet. If so, the caching mechanism is skipped for this
workload.

Change-Id: Ie7bdf2d863f433e7998b1b92cea129c6bda4dd8f