Schedulable archs flag (#68)
* Add `cluster-schedulable-archs` flag

 # Context

In some cases we have seen that the common architectures of the images
do not match the default preferred architecture. In those cases, one
of the common architectures is selected, which may not match the
available architectures in the cluster. In those cases, the Pod is
always pending scheduling waiting for a node that will never exist.

 # Approach

Allow passing a comma-separated list of architectures schedulable in
the cluster. As we detect image's architecture support, we ignore
those architectures that have not been specified as supported by the
cluster.

This new flag is optional and defaults to empty string (which disables
the feature and keeps the prior behaviour).

 ## Startup fail if preferred architecture is not part of the supported architectures

Startup will fail if the preferred architecture selected via
commandline is not part of the list of schedulable architectures to
prevent misconfigurations.

 ## Ignore preferred arch at Pod level if unsupported

If a preferred architecture is indicated at Pod level, but it does not
belong to the list of schedulable architectures, we ignore the preference
and treat it as not specified. A log line is generated in that case.

 # Non-goal

 ## Detect architectures present in the cluster

We do not want to prevent scheduling of Pods with architectures
supported but which have no nodes currently in the cluster
(downscaled) to avoid interfering with scale from 0 set ups.

Change-Id: I0b85ab61fb3a267b41d938bd90f2e2d2050e5447
Co-authored-by: Fabián Sellés Rosa <1088313+Fsero@users.noreply.github.com>

* fixup! Add `cluster-schedulable-archs` flag

* fixup! Add `cluster-schedulable-archs` flag

---------

Co-authored-by: Fabián Sellés Rosa <1088313+Fsero@users.noreply.github.com>