Refactor registry client, extract listArchsWithAuth to dedicated function
We have recently seen that in case an error occurs when getting a single
manifest fails.
In those cases, it could happen that the ListArch func drops some of the
supported architecture, randomly.

Take a step forward towards solving this issue. Isolate retrieving image
architectures for a single authorisation candidate

Change-Id: I7f9e27cff2faddad54145a84cb677458cdc5f6c6