Add missing image manifest header
Problem
---

When retrieving image distribution manifest list, we send the request
without any Accept header.

This causes some image registries to return 404, and makes us miss the
real image architectures supported.

This PR ensures that we are able to to fetch individual image manifests
so we can successfully detect the supported architectures for most of
the registries

Change-Id: I4d1d16076148659067a2025a0c4fe54c3fe26d6f