Refactor Authenticator to inject the channel
Context
---

We currently suport 2 authentication sources. Docker and containerD.

Problems
---

1. Kubelet has a third one: [CredentialProviderConfig]

To be consistent, noe needs to implement all of those to be able to
lookup the values in the relevant place.

2. The current containerD authentication is mixed with the Docker one (file
and image pull secret ones).

Goal
---

Refactor the authentication provider in order to split them into
one provider by kind of authentication.

This will ease imlement the kubelet authentication and manage the
different authentication mechanisms

[CredentialProviderConfig]: https://kubernetes.io/docs/tasks/administer-cluster/kubelet-credential-provider/

Change-Id: Ia9abeec4a731b22613148daa634f49b8604990e0
Signed-off-by: Thibault Jamet <thibault.jamet@adevinta.com>