Improve code contributions guidelines
As we have open sourced noe and opened contributions, we observe the
need of adjusting what are the internal contributions guidelines.

Because noe is now an opensource project, we need to ensure that our
contribution recommendation respects contributors and company privacies
and that we reflect our intentions with code reviews.

Change-Id: I003cdb0eff8d3801f330fb6a907d53da3fa2ee53