Extract cache logic from the cache registry
Context
---

We are observing an important number of 429 (too many requests) do
Dockerhub.

This is due to the fact that we authenticate to DockerHub for each
request of each manifest.

In some cases, images are built for many architectures, like the alpine one
which is built for 7 architectures.

This plays against the [DockerHub rate limits](https://www.docker.com/increase-rate-limits/)
and consume too much of those (from a 100 to 5k pulls per day depending
on the users)

Goal
---

Refactor the code so the cache is easily re-usable for the
authentication mechanisms

Change-Id: Idb0db1cdf9ffce117f6783975f5dfda794e46cf9