Add cache to the www-authenticate roundtripper
Context
---

We are observing an important number of 429 (too many requests) do
Dockerhub.

This is due to the fact that we authenticate to DockerHub for each
request of each manifest.

In some cases, images are built for many architectures, like the alpine one
which is built for 7 architectures.

This plays against the [DockerHub rate limits](https://www.docker.com/increase-rate-limits/)
and consume too much of those (from a 100 to 5k pulls per day depending
on the users)

Goal
---

Reduce the number of authentication to DockerHub as much as possible

Change-Id: I5e5b22fe14aa8da18f574ba40f68985a3710b02d