Fix latest tag not being built
When building a specific sha, a pull is performed to avoid re-building existing images
and ensure image immutability.

When building the latest image, we must ensure that we always provide the latest build as, by design,
there will always be an existing image with the latest tag.
Latest images being mutable

Fixes: #107

Change-Id: Ib4fa955a0b802764b495aebf527650f1d3c35cb6