Fix ListArchs return values.
Context:

Due to some third-party errors, such as issues with the registry, the listArchs function may not return an error but instead return an empty platform.

To address this, I implemented a break in the loop to exit once we successfully list architectures and I handle the return values after the loop.